requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
textblob==0.15.3
google-api-python-client==2.47.0
python-dotenv==0.19.1

